are you kidding me

+++ runge.cu
@@ -212,7 +212,7 @@ supplies the routine derivs(x,y,dydx) , which returns derivatives dydx at x .
 void rk4(SphVector y_d[], SphVector dydx_d[], int n, double x, double h, SphVector yout[], void (*derivs)(double, SphVector[], SphVector[], int, Vector[]), bool CopyToHost) {
 	double xh, hh, h6; 
 	dim3 blockDim(BLOCK_SIZE, BLOCK_SIZE, BLOCK_SIZE);
-        dim3 gridDim(ceil(WIDTH/BLOCK_SIZE), ceil(HEIGHT/BLOCK_SIZE), ceil(DEPTH/BLOCK_SIZE));	
+        dim3 gridDim(ceil((float)WIDTH/(float)BLOCK_SIZE), ceil((float)HEIGHT/(float)BLOCK_SIZE), ceil((float)DEPTH/(float)BLOCK_SIZE));	
 
 	//device arrays
 	SphVector *dym_d, *dyt_d, *yt_d;
@@ -311,7 +311,7 @@ and xx[0..nstep].
 void rkdumb(SphVector vstart[], int nvar, double x1, double x2, int nstep, void (*derivs)(double, SphVector[], SphVector[], int, Vector[])) {
 	double x, h;
 	dim3 blockDim(BLOCK_SIZE, BLOCK_SIZE, BLOCK_SIZE);
-        dim3 gridDim(ceil(WIDTH/BLOCK_SIZE), ceil(HEIGHT/BLOCK_SIZE), ceil(DEPTH/BLOCK_SIZE));	
+        dim3 gridDim(ceil((float)WIDTH/(float)BLOCK_SIZE), ceil((float)HEIGHT/(float)BLOCK_SIZE), ceil((float)DEPTH/(float)BLOCK_SIZE));	
 	SphVector *v, *vout, *dv;
 
 	//device arrays
